Create array-sum.php
When I put in this information, the code is passing but there is a PHP Notice:

PHP Notice:  Undefined offset: 6 in array-sum.php on line 15

I know that this means that my array is not being seen by the page, but I don't know how to fix it.  Since the code is passing, I am worried I am learning the wrong way to do this. If you might know why this isn't working, please let me know.